requirements.txt
Flask==2.2.2
google-cloud-datastore==2.7.1
google-auth==2.16.0
requests==2.28.2
Flask-Caching==2.0.2
